Novoto

+++ TanyaTeamWork/Markets.Model/Product.cs
@@ -1,5 +1,6 @@
 ﻿namespace Markets.Model
 {
+    //OKI
     using System.ComponentModel.DataAnnotations;
 
     public class Product
@@ -9,7 +10,7 @@
         [MaxLength(50)]
         public string ProductName { get; set; }
 
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
         public int VendorId { get; set; }
         public virtual Vendor Vendors { get; set; }
